Move LogicalPlan to plan.rs
